+++ src/store/configStore.js
@@ -1,6 +1,7 @@
-import { createStore } from 'redux';
+import { createStore, applyMiddleware } from 'redux';
 import combineReducer from './combineReducer';
+import thunk from 'redux-thunk'
 
-const store = createStore(combineReducer);
+const store = createStore(combineReducer, applyMiddleware(thunk));
 
 export default store