+++ src/store/time/middleware.js
@@ -1,0 +1,9 @@
+import { getTime } from './actions';
+
+export function fetchTime(){
+	return (dispatch) => {
+		setInterval(() => {
+			dispatch(getTime({ now: new Date().getTime() }))
+        }, 1000);
+	}
+}