+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 ### 20190812  初始化react-cli(配置react + react-router + react-redux)
 
-### 20190814  添加css module配置，普通模式和css module模式同时兼容
+### 20190814  
+1. 添加css module配置，普通模式和css module模式同时兼容  
+2. 添加redux-thunk